backend/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
